Renamed LSP related files of the grammar language
