+++ packages/langium/src/grammar/lsp/grammar-folding-ranges.ts
