+++ packages/langium/src/grammar/lsp/grammar-formatter.ts
