+++ packages/langium/src/grammar/lsp/grammar-goto.ts
