+++ packages/langium/src/grammar/lsp/grammar-hovers.ts
